+++ lib/routes.jsx
@@ -33,9 +33,11 @@ var pages = {
   'community/curriculum-workshop': require('../pages/curriculum-workshop.jsx'),
   'community/curriculum-workshop/march-8-2016': require('../pages/curriculum-workshop-march-8-2016.jsx'),
   'community/curriculum-workshop/april-12-2016': require('../pages/curriculum-workshop-april-12-2016.jsx'),
+  'community/curriculum-workshop/may-10-2016': require('../pages/curriculum-workshop-may-10-2016.jsx'),
   'community/community-call': require('../pages/community-call.jsx'),
   'community/community-call/march-23-2016': require('../pages/community-call-march-23.jsx'),
   'community/community-call/april-20-2016': require('../pages/community-call-april-20.jsx'),
+  'community/community-call/may-25-2016': require('../pages/community-call-may-25.jsx'),
   // NOTE: 'encryption' is reserved. See https://github.com/mozilla/teach.mozilla.org/issues/1798
   'events': require('../pages/events.jsx'),
   'events/resources': require('../pages/event-resources.jsx'),