+++ pages/community-call.jsx
@@ -28,19 +28,19 @@ var CommunityCallPage = React.createClass({
 
           <section className="callout-box">
             <h2>Upcoming Call</h2>
-            <p className="date">May 25th, 11am HADT / 1pm PT / 4pm ET / 8pm GMT / 10pm SAST</p>
-            <h1>Gigabit</h1>
+            <p className="date">June 29, 1pm PT / 4pm ET / 8pm UTC</p>
+            <h1>Learning and Making</h1>
             <p className="description">
-              Featured speakers: Caleb Bagby, teacher at Red Bank High School in Chattanooga; Chad Sansing, Curriculum Developer and Web Literacy for Mozilla; and Rebecca Dove and Quest Taylor, who are working together on an artificial intelligence literacy application called Pennez.
+              Featured Speakers: Lisa Kim, Youth Commission Director for Mikva Challenge in Chicago, IL; Niberca (Gigi) Polo, Emoti-con 2016 design fellows lead, Part-Time Associate Teaching Professor at Parsons The New School for Design, and Principal at Myellow Boots Studio; Dr. Dixie Ching, co-lead of the Hive Research Lab; Andy Forest, lead at STEAMLabs.
             </p>
           </section>
 
           <p>
-            Technology is continually advancing. Join us as we explore how high-speed, low-latency gigabit networks are allowing educational technology to advance rapidly. Thanks to these new connection speeds, education has the opportunity to evolve in ways that wouldn’t be possible on traditional networks, becoming more immersive and more engaging than ever before.
+            Summer learning loss is a real challenge, especially for those who may not have access to enriching educational opportunities. There has been a good amount of research and discussion on this topic, and on the June community call, we’ll be exploring the types of fun and engaging learning experiences that can–and do–happen during the summer. Educators from Hive NYC, Hive Chicago and Hive Toronto will join us to share real examples of programs that illustrate the impact that learning and making can have on young people during this critical time of the year.
           </p>
 
           <p>
-            On this month’s call, we’ll explore how gigabit technology is transforming today’s classroom. We’ll discuss how emerging technologies like virtual reality, artificial intelligence, and 4K video streaming are being deployed to engage students, address learning needs, and create new outcomes for education.
+            How do these summer programs Demystify the Web for young learners? Explore our learning space at the Mozilla Festival and how you can submit a proposal to share your program strategies and best practices with a community of like-minded peers in a professional learning environment that models hands-on, interest-driven learning for youth and adults alike.
           </p>
 
           <p>
@@ -50,32 +50,39 @@ var CommunityCallPage = React.createClass({
           <h3>Workshop Video Stream</h3>
 
           <div className="video-wrapper">
-            <iframe width="560" height="315" src="//www.youtube.com/embed/LlEXY8NKxoY" frameborder="0" allowfullscreen></iframe>
+            <iframe width="560" height="315" src="//www.youtube.com/embed/Uig7s0fChj8" frameBorder="0" allowFullScreen></iframe>
           </div>
 
           <h4>
             Open Agenda
-            <a title="Open the agenda in a new tab" className="fa fa-external-link open-etherpad" href="https://public.etherpad-mozilla.org/p/mozcommunitycallMay16">
+            <a title="Open the agenda in a new tab" className="fa fa-external-link open-etherpad" href="https://public.etherpad-mozilla.org/p/mozcommunitycallJune16">
             </a>
           </h4>
 
-          <iframe className="etherpad" src="https://public.etherpad-mozilla.org/p/mozcommunitycallMay16"></iframe>
+          <iframe className="etherpad" src="https://public.etherpad-mozilla.org/p/mozcommunitycallJune16"></iframe>
 
           <h2>Upcoming Calls</h2>
 
           <ul className="upcoming-workshops">
             <li>
-              <p className="date">June 29th</p>
-              <h2>Theme: Summer Learning</h2>
-              <p>
-                 Featured Speakers: TBD
-              </p>
+              <p className="date">July TBD</p>
+              <h2>Theme: TBD</h2>
             </li>
           </ul>
 
           <h2>Past Calls</h2>
 
           <ul className="past-workshops">
+            <li>
+              <p className="date">May 25, 2016</p>
+              <h2>Gigabit</h2>
+              <p>
+                On this month’s call, we’ll explore how gigabit technology is transforming today’s classroom. We’ll discuss how emerging technologies like virtual reality, artificial intelligence, and 4K video streaming are being deployed to engage students, address learning needs, and create new outcomes for education.
+              </p>
+              <p className="watch-archive">
+                <LinkAnchorSwap to="/community/community-call/may-25-2016/">Watch the Replay</LinkAnchorSwap>
+              </p>
+            </li>
             <li>
               <p className="date">April 20, 2016</p>
               <h2>Internet of Things</h2>